+++ packages/powersync_core/test/watch_test.dart
@@ -53,7 +53,7 @@ void main() {
           'INSERT INTO customers(id, name) VALUES (?, ?)', [id, 'a customer']);
 
       var done = false;
-      inserts() async {
+      Future<void> inserts() async {
         while (!done) {
           await powersync.execute(
               'INSERT INTO assets(id, make, customer_id) VALUES (uuid(), ?, ?)',
@@ -65,7 +65,7 @@ void main() {
 
       const numberOfQueries = 10;
 
-      inserts();
+      final insertsFuture = inserts();
       try {
         List<DateTime> times = [];
         final results = await stream.take(numberOfQueries).map((e) {
@@ -100,6 +100,8 @@ void main() {
       } finally {
         done = true;
       }
+
+      await insertsFuture;
     });
 
     test('onChange', () async {
@@ -111,7 +113,7 @@ void main() {
       const throttleDuration = Duration(milliseconds: baseTime);
 
       var done = false;
-      inserts() async {
+      Future<void> inserts() async {
         while (!done) {
           await powersync.execute(
               'INSERT INTO assets(id, make) VALUES (uuid(), ?)', ['test']);
@@ -120,7 +122,7 @@ void main() {
         }
       }
 
-      inserts();
+      final insertsFuture = inserts();
 
       final stream = powersync.onChange({'assets', 'customers'},
           throttle: throttleDuration).asyncMap((event) async {
@@ -138,6 +140,7 @@ void main() {
             UpdateNotification.single('assets'),
             UpdateNotification.single('assets')
           ]));
+      await insertsFuture;
     });
 
     test('emits update events with friendly names', () async {