+++ packages/powersync_attachments_helper/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.6.18+11
+
+ - Update a dependency to the latest release.
+
+## 0.6.18+10
+
+ - Update a dependency to the latest release.
+
 ## 0.6.18+9
 
  - Update a dependency to the latest release.